backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
scikit-learn>=1.5.0
numpy>=2.0.0
joblib>=1.4.2
pydantic>=2.7.1
python-multipart>=0.0.9
python-pptx>=0.6.23
PyPDF2>=3.0.1
